Final clean ups
