--- app/src/main/java/android/hardware/fingerprint/MyAuthenticationResult.kt
@@ -1,20 +1,0 @@
-@file:Suppress("DEPRECATION")
-
-package android.hardware.fingerprint
-
-import android.hardware.biometrics.BiometricPrompt
-
-class MyAuthenticationResult(private val result: BiometricPrompt.AuthenticationResult) :
-    FingerprintManager.AuthenticationResult() {
-
-    private fun BiometricPrompt.CryptoObject.toLegacy() = when {
-        cipher != null -> FingerprintManager.CryptoObject(cipher)
-        mac != null -> FingerprintManager.CryptoObject(mac)
-        signature != null -> FingerprintManager.CryptoObject(signature)
-        else -> null
-    }
-
-    override fun getCryptoObject(): FingerprintManager.CryptoObject? {
-        return result.cryptoObject?.toLegacy()
-    }
-}
