+++ app/src/main/java/android/hardware/fingerprint/MyFingerprintManager.kt
@@ -8,7 +8,6 @@ import android.hardware.biometrics.BiometricManager
 import android.hardware.biometrics.BiometricPrompt
 import android.os.CancellationSignal
 import android.os.Handler
-import android.util.Log
 import com.edison.fingerface.HandlerExecutor
 import com.edison.fingerface.MainActivity
 import com.edison.fingerface.PreferenceProvider
@@ -19,7 +18,6 @@ class MyFingerprintManager(private val context: Context) : FingerprintManager() 
     private val prefs = PreferenceProvider.getRemote(context)
 
     override fun isHardwareDetected(): Boolean {
-        Log.d(TAG, "isHardwareDetected")
         return when (manager.canAuthenticate()) {
             BiometricManager.BIOMETRIC_ERROR_NONE_ENROLLED,
             BiometricManager.BIOMETRIC_SUCCESS -> true
@@ -28,7 +26,6 @@ class MyFingerprintManager(private val context: Context) : FingerprintManager() 
     }
 
     override fun hasEnrolledFingerprints(): Boolean {
-        Log.d(TAG, "hasEnrolledFingerprints")
         return when (manager.canAuthenticate()) {
             BiometricManager.BIOMETRIC_SUCCESS -> true
             else -> false
@@ -46,8 +43,6 @@ class MyFingerprintManager(private val context: Context) : FingerprintManager() 
         crypto: CryptoObject?, cancel: CancellationSignal?,
         flags: Int, legacyCallback: AuthenticationCallback, handler: Handler?
     ) {
-        Log.d(TAG, "authenticate")
-
         val bioCrypto = crypto?.toBio()
         val cancelSignal = cancel ?: CancellationSignal()
         val executor = handler?.let { HandlerExecutor(it) } ?: context.mainExecutor
@@ -68,7 +63,40 @@ class MyFingerprintManager(private val context: Context) : FingerprintManager() 
         }
     }
 
-    companion object {
-        const val TAG = "Fingerface"
+    class MyAuthenticationCallback(
+        private val callback: AuthenticationCallback
+    ) : BiometricPrompt.AuthenticationCallback() {
+
+        override fun onAuthenticationError(errorCode: Int, errString: CharSequence) {
+            callback.onAuthenticationError(errorCode, errString)
+        }
+
+        override fun onAuthenticationHelp(helpCode: Int, helpString: CharSequence) {
+            callback.onAuthenticationHelp(helpCode, helpString)
+        }
+
+        override fun onAuthenticationFailed() {
+            callback.onAuthenticationFailed()
+        }
+
+        override fun onAuthenticationSucceeded(result: BiometricPrompt.AuthenticationResult) {
+            callback.onAuthenticationSucceeded(MyAuthenticationResult(result))
+        }
+    }
+
+    class MyAuthenticationResult(
+        private val result: BiometricPrompt.AuthenticationResult
+    ) : AuthenticationResult() {
+
+        private fun BiometricPrompt.CryptoObject.toLegacy() = when {
+            cipher != null -> CryptoObject(cipher)
+            mac != null -> CryptoObject(mac)
+            signature != null -> CryptoObject(signature)
+            else -> null
+        }
+
+        override fun getCryptoObject(): CryptoObject? {
+            return result.cryptoObject?.toLegacy()
+        }
     }
 }
